+++ src/main/java/se/ifmo/address/AddressRepository.java
@@ -4,5 +4,5 @@ import org.springframework.stereotype.Repository;
 import se.ifmo.common.placemark.AbstractRepository;
 
 @Repository
-public interface AddressRepository extends AbstractRepository<Address, Long> {
+public interface AddressRepository extends AbstractRepository<Address, Integer> {
 }
