+++ src/main/java/se/ifmo/coordinates/CoordinatesRepository.java
@@ -4,5 +4,5 @@ import org.springframework.stereotype.Repository;
 import se.ifmo.common.placemark.AbstractRepository;
 
 @Repository
-public interface CoordinatesRepository extends AbstractRepository<Coordinates, Long> {
+public interface CoordinatesRepository extends AbstractRepository<Coordinates, Integer> {
 }
