+++ src/main/java/se/ifmo/location/LocationRepository.java
@@ -4,5 +4,5 @@ import org.springframework.stereotype.Repository;
 import se.ifmo.common.placemark.AbstractRepository;
 
 @Repository
-public interface LocationRepository extends AbstractRepository<Location, Long> {
+public interface LocationRepository extends AbstractRepository<Location, Integer> {
 }
